refactor: move platform settings func to views
