--- lms/djangoapps/learner_dashboard/learner_api.py
@@ -1,15 +1,0 @@
-"""APIs used to source data/interactions for the Learner Dashboard"""
-
-from django.conf import settings
-
-from common.djangoapps.edxmako.shortcuts import marketing_link
-
-
-def get_platform_settings():
-    """Get settings used for platform level connections: emails, url routes, etc."""
-
-    return {
-        "supportEmail": settings.DEFAULT_FEEDBACK_EMAIL,
-        "billingEmail": settings.PAYMENT_SUPPORT_EMAIL,
-        "courseSearchUrl": marketing_link("COURSES"),
-    }
