+++ lms/djangoapps/learner_dashboard/test_learner_views.py
@@ -1,16 +1,48 @@
-"""Test for learner views"""
+"""Test for learner views and related functions"""
 
 import json
+from unittest import TestCase
+from unittest.mock import patch
+from uuid import uuid4
 
 from django.urls import reverse
 from rest_framework.test import APITestCase
 
+from lms.djangoapps.learner_dashboard.learner_views import get_platform_settings
 from common.djangoapps.student.tests.factories import UserFactory
 from xmodule.modulestore.tests.django_utils import (
     TEST_DATA_SPLIT_MODULESTORE,
     SharedModuleStoreTestCase,
 )
 from xmodule.modulestore.tests.factories import CourseFactory
+
+
+class TestGetPlatformSettings(TestCase):
+    """Tests for get_platform_settings"""
+
+    MOCK_SETTINGS = {
+        "DEFAULT_FEEDBACK_EMAIL": f"{uuid4()}@example.com",
+        "PAYMENT_SUPPORT_EMAIL": f"{uuid4()}@example.com",
+    }
+
+    @patch.multiple("django.conf.settings", **MOCK_SETTINGS)
+    @patch("lms.djangoapps.learner_dashboard.learner_views.marketing_link")
+    def test_happy_path(self, mock_marketing_link):
+        # Given email/search info exists
+        mock_marketing_link.return_value = mock_search_url = f"/{uuid4()}"
+
+        # When I request those settings
+        return_data = get_platform_settings()
+
+        # Then I return them in the appropriate format
+        self.assertDictEqual(
+            return_data,
+            {
+                "supportEmail": self.MOCK_SETTINGS["DEFAULT_FEEDBACK_EMAIL"],
+                "billingEmail": self.MOCK_SETTINGS["PAYMENT_SUPPORT_EMAIL"],
+                "courseSearchUrl": mock_search_url,
+            },
+        )
 
 
 class TestDashboardView(SharedModuleStoreTestCase, APITestCase):
